[hotfix][connectors/elasticsearch] Move files from streaming.connectors.elasticsearch.table to connector.elasticsearch.table
