--- flink-connectors/flink-connector-elasticsearch6/src/main/java/org/apache/flink/streaming/connectors/elasticsearch/table/Elasticsearch6ConnectorOptions.java
@@ -1,39 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.flink.streaming.connectors.elasticsearch.table;
-
-import org.apache.flink.annotation.PublicEvolving;
-import org.apache.flink.configuration.ConfigOption;
-import org.apache.flink.configuration.ConfigOptions;
-
-/**
- * Options specific for the Elasticsearch 6 connector. Public so that the {@link
- * org.apache.flink.table.api.TableDescriptor} can access it.
- */
-@PublicEvolving
-public final class Elasticsearch6ConnectorOptions extends ElasticsearchConnectorOptions {
-
-    private Elasticsearch6ConnectorOptions() {}
-
-    public static final ConfigOption<String> DOCUMENT_TYPE_OPTION =
-            ConfigOptions.key("document-type")
-                    .stringType()
-                    .noDefaultValue()
-                    .withDescription("Elasticsearch document type.");
-}
