+++ flink-connectors/flink-connector-elasticsearch7/src/main/resources/META-INF/services/org.apache.flink.table.factories.Factory
@@ -13,4 +13,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-org.apache.flink.streaming.connectors.elasticsearch.table.Elasticsearch7DynamicSinkFactory
+org.apache.flink.connector.elasticsearch.table.Elasticsearch7DynamicSinkFactory
